(refactor, movie list): Update naming and SOC
- separate logic from UI
- use sensible variable names in accordance to clean code
